Chapter 8/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
asyncpg==0.30.0
deprecation==2.1.0
iso8601==2.1.0
lancedb==0.21.2
numpy==2.2.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pyarrow==19.0.1
pydantic==2.11.3
pydantic_core==2.33.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tortoise-orm==0.25.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
